requirements.txt
et-xmlfile==1.1.0
openpyxl==3.1.3
requests==2.32.3
python-dotenv==1.0.1
wheel @ file:///usr/local/Cellar/python%403.12/3.12.3/libexec/wheel-0.43.0-py3-none-any.whl#sha256=80a49545e5e5ef2abeb79ec254447300d41dfb76914fc65efa6da8dd0a766052
